agentic_rag/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
PyYAML==6.0.2
llama-stack-client
fire
